requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Caching==2.1.0
Flask-JWT-Extended==4.5.3
Flask-Bcrypt==1.0.1
GeoAlchemy2==0.14.4
Shapely==2.0.2  # For geometry operations
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.7  # For PostgreSQL connection
redis==5.0.1  # For Redis cache
numpy<2.0
